fix(investments): give positions split its own taller default with chart
`PositionsTransactionsSplit` was sharing a single `NSSplitView` autosave
key and a 180pt initial top height across all call sites. That works for
the chartless multi-currency case, but for an investment account in
`.calculatedFromTrades` mode `PositionsView` adds a 220pt fixed-height
chart between header and table — at 180pt the chart pushes the
positions table off-screen so the user appears to see only "value graph
+ transaction list" until they drag the divider down.

Add `autosaveName` and `initialTopHeight` parameters to
`PositionsTransactionsSplit` (defaults preserved for the chartless call
site) and have `InvestmentAccountView.positionTrackedLayout` opt into a
distinct key (`positions-transactions-split.with-chart`) and a 540pt
initial top height so header + chart + several table rows render without
intervention. Distinct keys also stop the saved divider from one layout
bleeding into the other.

Also hoists the previews + seed helpers into
`InvestmentAccountView+Previews.swift` to keep the main view file under
its 400-line budget.
